added varied array modal component
