+++ src/components/index.js
@@ -1,10 +1,12 @@
-import * as ArrayFields from './ArrayFields';
+import * as UniformedArrayFields from './UniformedArrayFields';
+import * as VariedArrayFields from './VariedArrayFields';
 import * as InputFields from './InputFields';
 import * as CardContainers from './CardContainers';
 import * as SelectFields from './SelectFields';
 import * as ListItems from './ListItems';
 
-export * from './ArrayFields';
+export * from './UniformedArrayFields';
+export * from './VariedArrayFields';
 export * from './InputFields';
 export * from './CardContainers';
 export * from './SelectFields';
@@ -15,12 +17,13 @@ export FormField from './FormField';
 export const defaultComponents = {
   string: InputFields.Text,
   number: InputFields.Number,
-  array: ArrayFields.UniformedArrayInline
+  array: UniformedArrayFields.UniformedArrayInline
 };
 
 export default {
   ...InputFields,
-  ...ArrayFields,
+  ...UniformedArrayFields,
+  ...VariedArrayFields,
   ...CardContainers,
   ...SelectFields,
   ...ListItems
